requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==4.2.17
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
joblib==1.4.2
numpy==2.2.0
pandas==2.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.14.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websockets==14.1
zope.interface==7.2
